[screensaver.turnoff] v0.9.1
When fixing the xbmcgui.Dialog().notification() call I introduced a
small issue wrt. message. So every pop-up (on error) is now causing an
error. (The original v0.9.0 did not have this issue)

So I decided to do a quick v0.9.1 release as this could be annoying for
users testing the different methods.